+++ Codice/Code/Games/pong.h
@@ -34,10 +34,10 @@
 * [ TYPEDEFS STRUCTS ]
 *******************************************************************************/
 typedef struct Ball {
-    int x;
-    int y;
-    int xVel;
-    int yVel;
+    uint8_t x;
+    uint8_t y;
+    int8_t xVel;
+    int8_t yVel;
 } Ball;
 
 /******************************************************************************
@@ -46,20 +46,20 @@ typedef struct Ball {
 // init functions
 void initBall(Ball* ballPtr); // initializes the ball
 // moving/updating functions
-void moveUser(int* userUpperYPtr, int* lastUserYPtr); // moves the user
-void moveEnemy(int* enemyUpperYPtr, int* lastEnemyYPtr, Ball ball); // moves the enemy
+void moveUser(uint8_t* userUpperYPtr, uint8_t* lastUserYPtr); // moves the user
+void moveEnemy(uint8_t* enemyUpperYPtr, uint8_t* lastEnemyYPtr, Ball ball); // moves the enemy
 void moveBall(Ball* ballPtr); // moves the ball
 void invertBallXDirection(Ball* ballPtr); // inverts the ball horizontal direction
 void invertBallYDirection(Ball* ballPtr); // inverts the ball vertical direction
 void updateBallY(Ball* ballPtr); // updates the ball vertical velocity
 // bool functions
-bool checkUserCollision(int userUpperY, Ball ball); // checks if the ball collides with the user
-bool checkEnemyCollision(int enemyUpperY, Ball ball);  // checks if the ball collides with the enemy
+bool checkUserCollision(uint8_t userUpperY, Ball ball); // checks if the ball collides with the user
+bool checkEnemyCollision(uint8_t enemyUpperY, Ball ball);  // checks if the ball collides with the enemy
 bool checkPongBorderCollision(Ball ball); // checks if the ball collides with any border
 // draw functions
 void drawPongBackground(); // draws the background
-void drawPongUser(int userUpperY, int lastUserY); // draws the user
-void drawPongEnemy(int enemyUpperY, int lastEnemyY); // draws the enemy
+void drawPongUser(uint8_t userUpperY, uint8_t lastUserY); // draws the user
+void drawPongEnemy(uint8_t enemyUpperY, uint8_t lastEnemyY); // draws the enemy
 void drawBall(Ball ball); // draws the ball
 // run function
 void runPong(); // runs the Pong game
